% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.r
\name{address}
\alias{address}
\title{Call the C function address() that retrieves the memory address of an R object}
\usage{
address(x)
}
\arguments{
\item{x}{object whose memory address is of interest.}
}
\value{
the memory address of object \code{x} or \code{NULL} if the object does not exist in the R workspace.
}
\description{
Call the C function address() that retrieves the memory address of an R object
}
