% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_environment_name}
\alias{get_environment_name}
\title{Get the name of an environment when the address-name lookup table has not yet been constructed}
\usage{
get_environment_name(env)
}
\arguments{
\item{env}{environment whose name is of interest}
}
\value{
the name of the environment as given by \code{environmentName()} if it is a named environment
or the name of the environment as given by \code{deparse(substitute())} otherwise.
This will result in a string representing an expression when 'env' is given as an expression as in
\code{parent.env(env1)}.
}
\description{
Get the name of an environment when the address-name lookup table has not yet been constructed
}
\keyword{internal}
