% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_envs}
\alias{get_envs}
\title{Find environments in a given environment}
\usage{
get_envs(env_name, path, env_path_list, envir = .GlobalEnv)
}
\arguments{
\item{env_name}{name of the environment where the search for environments should be carride out.}

\item{path}{array with environment names leading to the \code{env_name} environment.}

\item{env_path_list}{array or list of environment names found so far including their path.}

\item{envir}{root environment where the search for environments started.}
}
\value{
Either an updated \code{env_path_list} with the environments found so far in the environments tree
or TRUE, which indicates that a leaf in the environments tree found inside the \code{envir} environment
has been reached.
}
\description{
Function that looks for the environments defined within a given environment
}
\keyword{internal}

