% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_namespace_names}
\alias{get_namespace_names}
\title{Return the names that can be used to load all the namespace environments in the search() list}
\usage{
get_namespace_names()
}
\value{
Array containing the names of the namespace environments as values and the package names as names
(in the form of e.g. "package:base"). The names of the namespace environments are the result of splitting
the name in the search list by ":" and taking the second part of the string 
(e.g. "base" from "package:base" or "rstudio" from "tools:rstudio").
The idea is that this name gives the namespace enviroment when used as \code{asNamespace(<name>)} or
\code{getNamespace(<name>)}, as in e.g. \code{asNamespace("base")}.
}
\description{
This function returns the names of all the namespace environments of the packages found
in the \code{search()} path.
}
\keyword{internal}
