% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_objects_in_package}
\alias{get_objects_in_package}
\title{Get the objects defined in a given package's namespace}
\usage{
get_objects_in_package(package_name)
}
\arguments{
\item{package_name}{string containing the package name (e.g. "envnames") of interest.}
}
\value{
An array containing the objects defined in a package namespace, as obtained
by \code{ls(asNamespace(package_name))}, or NULL if the package does not exist.
}
\description{
Get the objects defined in a given package's namespace
}
\keyword{internal}

