% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_searchpath_environment_addresses}
\alias{get_searchpath_environment_addresses}
\title{Return the system and package environments in the search path}
\usage{
get_searchpath_environment_addresses()
}
\value{
a named array containing the addresses of the system and package environments
found in the \code{search()} path with the environment names as its names attribute.
}
\description{
Return the system and package environments in the search path
}
\keyword{internal}
