% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_user_environment_names_in_env}
\alias{get_user_environment_names_in_env}
\title{Look for user environments defined inside an environment}
\usage{
get_user_environment_names_in_env(envir)
}
\arguments{
\item{envir}{environment where the search should be carried out.}
}
\value{
an array containing the names of the user environments found
in the \code{envir} environment or NULL if \code{envir} is not an environment.
}
\description{
Look for user environments defined inside an environment
}
\keyword{internal}
