% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_user_environment_names_in_search_path}
\alias{get_user_environment_names_in_search_path}
\title{Look for user environments defined inside any of the system/package environments
of the search path}
\usage{
get_user_environment_names_in_search_path()
}
\value{
an array containing the names of the user environments found
inside any of the system/package environments in the search path (retrieved by \code{search()}).
The names attribute of the array is the system/package environment where the
environment is found.
\code{NULL} is returned if the process fails or no environments are found.
}
\description{
Look for user environments defined inside any of the system/package environments
of the search path
}
\keyword{internal}
