% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{is_null_or_na}
\alias{is_null_or_na}
\title{Check whether an object is NULL or NA.}
\usage{
is_null_or_na(x)
}
\arguments{
\item{x}{object to check.}
}
\value{
boolean indicating whether the object is NULL or NA.
FALSE is returned if the object does not exist.
}
\description{
This function silently handles special cases for which is.null() and is.na() may return a warning,
such as functions objects or environments.
(e.g. the warning "Warning message: In is.na(g) : is.na() applied to non-(list or vector) of type 'closure')"
}
\keyword{internal}
