% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{is_string}
\alias{is_string}
\title{Check whether an object is a string.}
\usage{
is_string(x)
}
\arguments{
\item{x}{object to check.}
}
\value{
boolean indicating whether the object is a string.
}
\description{
WARNING: This function fails when the value of x is "x"!! (i.e. it returns TRUE even when object 'x' is NOT a string per se --i.e. it was not passed as "x")
}
\details{
The result of this function is different from is.character(x) since this function returns TRUE
for an array of character values!
}
\keyword{internal}

