% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{parse_memory_address}
\alias{parse_memory_address}
\title{Parse a string that represents a memory address}
\usage{
parse_memory_address(x)
}
\arguments{
\item{x}{string to parse.}
}
\value{
string containing the memory address represented by the input string after stripping any
extraneous pieces of string, namely : "0x" and "environment: " and after enclosing it in '<>'.
For 32-bit architecture the string would be of the form "<xxxxxxxx>" where x represents
digits between 0 and 9 and letters between "a" and "f". Ex: "<07830f40>"
For 64-bit architecture the string would be of the form "<xxxxxxxxxxxxxxxx>" where x represents
digits between 0 and 9 and letters between "a" and "f". Ex: "<07830f40>"
}
\description{
Parse a string representing a memory address so that the address is returned in the way it is stored
in the 'address' column of the data frame returned by get_env_names().
}
\keyword{internal}
