% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{reset_option_warn}
\alias{reset_option_warn}
\title{Resets the "warn" option to the value stored in the hidden variable .option_warn
(set in global_definitions.r to the original value of the "warn" option when the package is loaded).}
\usage{
reset_option_warn()
}
\description{
Resets the "warn" option to the value stored in the hidden variable .option_warn
(set in global_definitions.r to the original value of the "warn" option when the package is loaded).
}
\keyword{internal}
