% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{set_option_warn_to_nowarning}
\alias{set_option_warn_to_nowarning}
\title{Set the "warn" options to -1 to avoid warning messages. The hidden variable .option_warn
defined in global_definitions.r has already been set to the original value of the "warn" option,
at the moment when the package is loaded, so that we can reset it later.}
\usage{
set_option_warn_to_nowarning()
}
\description{
Set the "warn" options to -1 to avoid warning messages. The hidden variable .option_warn
defined in global_definitions.r has already been set to the original value of the "warn" option,
at the moment when the package is loaded, so that we can reset it later.
}
\keyword{internal}
