% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{standardize_env_name}
\alias{standardize_env_name}
\title{Standardize the name of an environment}
\usage{
standardize_env_name(env_name)
}
\arguments{
\item{env_name}{environment name to standardize.}
}
\value{
The name of the environment, where the global environment is represented as "R_GlobalEnv" and the
base environment is shown as "base".
}
\description{
This function standardizes the name of an environment so that environment names are consistent
with the output of base function \link{environmentName}.
It only has an effect when the environment is the global environment or the base environment,
which have different ways of referring to them, namely:
\code{globalenv()}, \code{.GlobalEnv}, \code{baseenv()}, \code{as.environment("package:base")}
}
\keyword{internal}

