% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_definitions.r
\docType{data}
\name{testenv}
\alias{testenv}
\title{Environment used in testing the package}
\format{An object of class \code{environment} of length 1.}
\usage{
testenv
}
\description{
Environment used in testing the package
}
\keyword{datasets}
