% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{unlist_with_names}
\alias{unlist_with_names}
\title{Call unlist and preserve the names}
\usage{
unlist_with_names(alist)
}
\arguments{
\item{alist}{list to unlist.}
}
\value{
Whenever possible, an array whose names attribute is set to the names of the elements of the list.
}
\description{
Function \code{unlist} is called so that the output is an array (whenever possible) whose names
attribute contains the names of the elements of the list (unchanged, i.e. without adding a number
to identify them as the regular \code{unlist()} function does).
}
\keyword{internal}

