% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{crawl_envs_in_env}
\alias{crawl_envs_in_env}
\title{Crawl environments in search of user-defined environments}
\usage{
crawl_envs_in_env(env)
}
\arguments{
\item{envs}{environment to search in. Note that these should be an environment, and NOT an environment name.}
}
\value{
An array containing strings that represent the path to each environment found inside the \code{env}
environment, and its sub-environments.
}
\description{
Crawl environments in search of user-defined environments
}
\keyword{internal}

